u32に変更

+++ include/others/util.hpp
@@ -172,10 +172,10 @@ struct extent final {
 };
 
 struct rgba final {
-    f32 r{0.f};
-    f32 g{0.f};
-    f32 b{0.f};
-    f32 a{255.f};
+    u32 r{0};
+    u32 g{0};
+    u32 b{0};
+    u32 a{255};
 
     [[nodiscard]] constexpr auto operator+() const noexcept -> rgba { return *this; }
 
@@ -186,22 +186,22 @@ struct rgba final {
     }
 };
 
-inline constexpr auto white   = rgba{.r = 255.f, .g = 255.f, .b = 255.f};
-inline constexpr auto silver  = rgba{.r = 192.f, .g = 192.f, .b = 192.f};
-inline constexpr auto gray    = rgba{.r = 128.f, .g = 128.f, .b = 128.f};
-inline constexpr auto black   = rgba{.r = 0.f, .g = 0.f, .b = 0.f};
-inline constexpr auto red     = rgba{.r = 255.f, .g = 0.f, .b = 0.f};
-inline constexpr auto maroon  = rgba{.r = 128.f, .g = 0.f, .b = 0.f};
-inline constexpr auto yellow  = rgba{.r = 255.f, .g = 255.f, .b = 0.f};
-inline constexpr auto olive   = rgba{.r = 128.f, .g = 128.f, .b = 0.f};
-inline constexpr auto lime    = rgba{.r = 0.f, .g = 255.f, .b = 0.f};
-inline constexpr auto green   = rgba{.r = 0.f, .g = 128.f, .b = 0.f};
-inline constexpr auto aqua    = rgba{.r = 0.f, .g = 255.f, .b = 255.f};
-inline constexpr auto teal    = rgba{.r = 0.f, .g = 128.f, .b = 128.f};
-inline constexpr auto blue    = rgba{.r = 0.f, .g = 0.f, .b = 255.f};
-inline constexpr auto navy    = rgba{.r = 0.f, .g = 0.f, .b = 128.f};
-inline constexpr auto fuchsia = rgba{.r = 255.f, .g = 0.f, .b = 255.f};
-inline constexpr auto purple  = rgba{.r = 128.f, .g = 0.f, .b = 128.f};
+inline constexpr auto white   = rgba{.r = 255, .g = 255, .b = 255};
+inline constexpr auto silver  = rgba{.r = 192, .g = 192, .b = 192};
+inline constexpr auto gray    = rgba{.r = 128, .g = 128, .b = 128};
+inline constexpr auto black   = rgba{.r = 0, .g = 0, .b = 0};
+inline constexpr auto red     = rgba{.r = 255, .g = 0, .b = 0};
+inline constexpr auto maroon  = rgba{.r = 128, .g = 0, .b = 0};
+inline constexpr auto yellow  = rgba{.r = 255, .g = 255, .b = 0};
+inline constexpr auto olive   = rgba{.r = 128, .g = 128, .b = 0};
+inline constexpr auto lime    = rgba{.r = 0, .g = 255, .b = 0};
+inline constexpr auto green   = rgba{.r = 0, .g = 128, .b = 0};
+inline constexpr auto aqua    = rgba{.r = 0, .g = 255, .b = 255};
+inline constexpr auto teal    = rgba{.r = 0, .g = 128, .b = 128};
+inline constexpr auto blue    = rgba{.r = 0, .g = 0, .b = 255};
+inline constexpr auto navy    = rgba{.r = 0, .g = 0, .b = 128};
+inline constexpr auto fuchsia = rgba{.r = 255, .g = 0, .b = 255};
+inline constexpr auto purple  = rgba{.r = 128, .g = 0, .b = 128};
 
 struct uv final {
     f32 u{0.f};
